requirements.txt
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==24.3.0
beautifulsoup4==4.12.3
Brotli @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_f7i0oxypt6/croot/brotli-split_1736182464088/work
bs4==0.0.2
cbor==1.0.0
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_d8j59rqun5/croot/certifi_1734473289913/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.19.0
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.6
filelock==3.16.1
FlagEmbedding==1.3.3
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.10
ijson==3.3.0
inscriptis==2.5.1
ir_datasets==0.5.9
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-openai==0.3.0
langchain-text-splitters==0.3.5
langsmith==0.2.10
lxml==5.3.0
lz4==4.3.3
MarkupSafe==3.0.2
marshmallow==3.25.1
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
openai==1.59.7
orjson==3.10.14
packaging==24.2
pandas==2.2.3
peft==0.14.0
pillow==11.1.0
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pyarrow==18.1.0
pyarrow-hotfix==0.6
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyparsing==3.2.1
PySocks @ file:///Users/ktietz/ci_310/pysocks_1643961536721/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.2
requests-toolbelt==1.0.0
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
sentence-transformers==3.3.1
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.41.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.19.1
torch==2.5.1
tqdm==4.67.1
transformers==4.44.2
trec-car-tools==2.6
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unlzw3==0.2.3
urllib3 @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_06_m8gdsy6/croot/urllib3_1727769822458/work
uvicorn==0.34.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
xxhash==3.5.0
yarl==1.18.3
zlib-state==0.1.9
